pyproject.toml
[build-system]
requires = ["setuptools>=63", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dbsync"
version = "0.1.0"
description = "Production-ready PostgreSQL schema synchronizer"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [ { name = "Mellettan" } ]
dependencies = [
    "loguru>=0.7",
    "psycopg[binary]>=3.3",
    "pytest>=8.1",
    "rich",
    "python-dotenv",
    "black"
]

[project.scripts]
dbsync = "dbsync.runner:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[project.urls]
Homepage = "https://example.com/dbsync"
